+++ jdplus-incubator-base/jdplus-advancedsa-base-parent/jdplus-advancedsa-base-core/src/main/java/module-info.java
@@ -12,4 +12,5 @@ module jdplus.advancedsa.base.core {
 
     exports jdplus.advancedsa.base.core.regarima;
     exports jdplus.advancedsa.base.core.movingtd;
+    exports jdplus.advancedsa.base.core.tarima;
 }