+++ jdplus-incubator-base/jdplus-highfreq-base-parent/jdplus-highfreq-base-core/src/main/java/jdplus/highfreq/base/core/regarima/HighFreqRegArimaModel.java
@@ -36,12 +36,14 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
+import jdplus.toolkit.base.api.timeseries.TsData;
+import jdplus.toolkit.base.api.timeseries.TsPeriod;
+import jdplus.toolkit.base.api.timeseries.TsResiduals;
 import jdplus.toolkit.base.core.arima.IArimaModel;
 import jdplus.toolkit.base.core.dstats.T;
 import jdplus.toolkit.base.core.math.matrices.FastMatrix;
 import jdplus.toolkit.base.core.modelling.GeneralLinearModel;
 import jdplus.toolkit.base.core.modelling.LightweightLinearModel;
-import jdplus.toolkit.base.core.modelling.Residuals;
 import jdplus.toolkit.base.core.modelling.regression.RegressionDesc;
 import jdplus.toolkit.base.core.regarima.RegArimaEstimation;
 import jdplus.toolkit.base.core.regarima.RegArimaModel;
@@ -167,11 +169,15 @@ public class HighFreqRegArimaModel<S extends IArimaModel, M extends ArimaDescrip
                 .period(period)
                 .hyperParametersCount(free)
                 .build();
-
-        Residuals residuals = Residuals.builder()
-                .type(ResidualsType.FullResiduals)
-                .res(fullRes)
-                .start(description.getEstimationDomain().getEndPeriod().plus(-fullRes.length()))
+        TsPeriod start=description.getEstimationDomain().getEndPeriod().plus(-fullRes.length());
+        TsResiduals residuals = TsResiduals.builder()
+                .type(ResidualsType.QR_Transformed)
+                .res(ll.e())
+                .ssq(ll.ssq())
+                .n(ll.dim())
+                .df(ll.degreesOfFreedom())
+                .dfc(ll.degreesOfFreedom()-free)
+                .tsres(TsData.of(start, fullRes))
                 .test(ResidualsDictionaries.MEAN, niid.meanTest())
                 .test(ResidualsDictionaries.SKEW, niid.skewness())
                 .test(ResidualsDictionaries.KURT, niid.kurtosis())
@@ -194,7 +200,7 @@ public class HighFreqRegArimaModel<S extends IArimaModel, M extends ArimaDescrip
 
     Estimation estimation;
 
-    Residuals residuals;
+    TsResiduals residuals;
 
     DoubleSeq independentResiduals;
     List<RegressionDesc> regressionItems;